# services/generation-service/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Generation Service API
  description: |
    Comprehensive API for AI content generation with advanced performance optimization,
    caching, monitoring, and analytics capabilities.

    ## Features
    - High-performance content generation
    - Intelligent caching strategies
    - Real-time monitoring and metrics
    - Resource optimization
    - Load balancing and scaling

    ## Performance Targets
    - Workflow execution: < 30 seconds
    - Concurrent workflows: 20 simultaneous
    - API response (cached): < 100ms
    - Memory limit: < 2GB
    - Cache hit ratio: > 70%
    - Success rate: > 95%
  version: 1.0.0
  contact:
    name: Generation Service Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/cache/analytics:
    get:
      tags:
        - Cache
      summary: Get cache analytics
      description: Returns cache performance analytics and insights
      operationId: getCacheAnalytics
      parameters:
        - name: period
          in: query
          description: Time period for analytics (1h, 24h, 7d, 30d)
          schema:
            type: string
            enum: [1h, 24h, 7d, 30d]
            default: 24h
      responses:
        '200':
          description: Cache analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheAnalytics'
  /api/cache/clear:
    post:
      tags:
        - Cache
      summary: Clear cache
      description: Clears cache entries based on specified criteria
      operationId: clearCache
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CacheClearRequest'
            example:
              cache_type: "prompt_result"
              pattern: "gpt-*"
              confirm: true
      responses:
        '200':
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheClearResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  # Performance Monitoring Endpoints
...
  /api/performance/analytics:
    get:
      tags:
        - Performance
      summary: Get performance analytics
      description: Returns performance analytics and trend data
      operationId: getPerformanceAnalytics
      parameters:
        - name: period
          in: query
          description: Time period for analytics
          schema:
            type: string
            enum: [1h, 24h, 7d, 30d]
            default: 24h
        - name: metrics
          in: query
          description: Specific metrics to include
          schema:
            type: array
            items:
              type: string
            example: ["response_time", "throughput", "error_rate"]
      responses:
        '200':
          description: Performance analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceAnalytics'
  /api/performance/optimize:
    post:
      tags:
        - Performance
      summary: Trigger performance optimization
      description: Manually trigger performance optimization processes
      operationId: optimizePerformance
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizationRequest'
            example:
              optimization_type: "memory"
              force: false
              parameters:
                target_memory_mb: 1024
      responses:
        '200':
          description: Optimization triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResponse'
        '400':
          description: Invalid optimization request
        '401':
          description: Unauthorized
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    CacheAnalytics:
      type: object
      properties:
        period:
          type: string
        trends:
          type: object
          properties:
            hit_ratio_trend:
              type: string
              enum: [increasing, decreasing, stable]
            usage_trend:
              type: string
              enum: [increasing, decreasing, stable]
        hottest_keys:
          type: array
          items:
            type: object
            properties:
              key_pattern:
                type: string
              access_count:
                type: integer
        recommendations:
          type: array
          items:
            type: string
    CacheClearRequest:
      type: object
      required:
        - confirm
      properties:
        cache_type:
          type: string
          description: Specific cache type to clear
        pattern:
          type: string
          description: Key pattern to match (supports wildcards)
        confirm:
          type: boolean
          description: Confirmation flag to prevent accidental clears
    CacheClearResponse:
      type: object
      properties:
        cleared_count:
          type: integer
        cache_type:
          type: string
        pattern:
          type: string
        timestamp:
          type: string
          format: date-time
...
    PerformanceAnalytics:
      type: object
      properties:
        period:
          type: string
        metrics:
          type: object
          properties:
            response_times:
              type: object
              properties:
                avg:
                  type: number
                p50:
                  type: number
                p95:
                  type: number
                p99:
                  type: number
            throughput:
              type: object
              properties:
                requests_per_second:
                  type: number
                peak_rps:
                  type: number
            error_rates:
              type: object
              properties:
                overall:
                  type: number
                by_endpoint:
                  type: object
                  additionalProperties:
                    type: number
        trends:
          type: object
          additionalProperties:
            type: string
    OptimizationRequest:
      type: object
      required:
        - optimization_type
      properties:
        optimization_type:
          type: string
          enum: [memory, cpu, cache, network, all]
        force:
          type: boolean
          default: false
        parameters:
          type: object
          additionalProperties: true
    OptimizationResponse:
      type: object
      properties:
        optimization_id:
          type: string
        status:
          type: string
          enum: [started, completed, failed]
        actions_taken:
          type: array
          items:
            type: string
        estimated_improvement:
          type: object
          properties:
            memory_saved_mb:
              type: number
            performance_gain_percent:
              type: number
        completion_time:
          type: string
          format: date-time
...